fix: courses.sort_order を実型どおり bigint として扱い 32bit へ絞らない
sqlc の型付けスキーマが courses.sort_order を integer と宣言していたが、
本番の実型は bigint（information_schema で確認）。この食い違いにより
INSERT のパラメータが int4 に落ち、int32 を超える値はエラーも出さずに
負数へ巻き戻って保存されていた（2147483648 → -2147483648 を実測）。

course_chapters / rich_documents と同じ欠陥で、それらは同 PR 内で修正済み。
宣言を実型に合わせ、キャストを ::bigint に、呼び出し側を int64 に揃える。

+++ backend/internal/adapter/persistence/course_repository.go
@@ -114,7 +114,7 @@ func (r *courseRepository) Create(ctx context.Context, c *domain.Course) error {
 		Description:     c.Description,
 		Category:        c.Category,
 		Language:        c.Language,
-		SortOrder:       int32(c.SortOrder), // 0 は SQL 側の COALESCE で既定 100 に倒す
+		SortOrder:       int64(c.SortOrder), // 0 は SQL 側の COALESCE で既定 100 に倒す
 		IsPublished:     c.IsPublished,
 		CreatedAt:       createdAt,
 		UpdatedAt:       updatedAt,
@@ -143,7 +143,7 @@ func (r *courseRepository) Update(ctx context.Context, c *domain.Course) error {
 		ID:          id64,
 		Title:       c.Title,
 		Description: c.Description,
-		SortOrder:   int32(c.SortOrder),
+		SortOrder:   int64(c.SortOrder),
 		IsPublished: c.IsPublished,
 	})
 	if err != nil {
