+++ backend/internal/adapter/persistence/queries/course.sql
@@ -24,7 +24,7 @@ VALUES (
   sqlc.arg(description),
   sqlc.arg(category),
   sqlc.arg(language),
-  COALESCE(NULLIF(sqlc.arg(sort_order)::int, 0), 100),
+  COALESCE(NULLIF(sqlc.arg(sort_order)::bigint, 0), 100),
   sqlc.arg(is_published),
   sqlc.arg(created_at),
   sqlc.arg(updated_at)
